Rename module: 'utils' -> 'cached', rename functions
